enhancement: moving ai endpoints to EE and conditionally add them if ai is enabled
