+++ packages/core/admin/ee/admin/src/services/ai.ts
@@ -1,20 +1,27 @@
 import { adminApi } from '../../../../admin/src/services/api';
-import { GetAiUsage } from '../../../../shared/contracts/admin';
+import { GetAiUsage, GetAiToken } from '../../../../shared/contracts/ai';
 
 const aiService = adminApi.injectEndpoints({
   endpoints: (builder) => ({
-    getAIUsage: builder.query<GetAiUsage.Response, GetAiUsage.Request['query']>({
-      query: (params) => ({
+    getAIUsage: builder.query<GetAiUsage.Response, void>({
+      query: () => ({
+        method: 'GET',
         url: `/admin/ai-usage`,
-        config: {
-          params,
-        },
       }),
+    }),
+    getAiToken: builder.query<GetAiToken.Response['data'], void>({
+      query: () => ({
+        method: 'GET',
+        url: '/admin/ai-token',
+      }),
+      transformResponse(res: GetAiToken.Response) {
+        return res.data;
+      },
     }),
   }),
   overrideExisting: false,
 });
 
-const { useGetAIUsageQuery } = aiService;
+const { useGetAIUsageQuery, useGetAiTokenQuery, useLazyGetAiTokenQuery } = aiService;
 
-export { useGetAIUsageQuery };
+export { useGetAIUsageQuery, useGetAiTokenQuery, useLazyGetAiTokenQuery };
