+++ packages/core/admin/ee/server/src/index.ts
@@ -10,6 +10,8 @@ import auditLogsController from './audit-logs/controllers/audit-logs';
 import { createAuditLogsService } from './audit-logs/services/audit-logs';
 import { createAuditLogsLifecycleService } from './audit-logs/services/lifecycles';
 import { auditLog } from './audit-logs/content-types/audit-log';
+import aiRoutes from './ai/routes/ai';
+import aiController from './ai/controllers/ai';
 import type { Core } from '@strapi/types';
 
 const getAdminEE = () => {
@@ -26,24 +28,28 @@ const getAdminEE = () => {
     controllers,
     routes,
   };
-  // Only add the other audit-logs APIs if the feature is enabled by the user and the license
-  if (
+
+  // TODO: Check if ai feature is enabled in EE license when it's available
+  const isAIEnabled = strapi.config.get('admin.ai.enabled', true);
+  const isAuditLogsEnabled =
     strapi.config.get('admin.auditLogs.enabled', true) &&
-    strapi.ee.features.isEnabled('audit-logs')
-  ) {
-    return {
-      ...eeAdmin,
-      controllers: {
-        ...eeAdmin.controllers,
-        'audit-logs': auditLogsController,
-      },
-      routes: {
-        ...eeAdmin.routes,
-        'audit-logs': auditLogsRoutes,
-      },
-      async register({ strapi }: { strapi: Core.Strapi }) {
-        // Run the the default registration
-        await eeAdmin.register({ strapi });
+    strapi.ee.features.isEnabled('audit-logs');
+  return {
+    ...eeAdmin,
+    controllers: {
+      ...eeAdmin.controllers,
+      ...(isAuditLogsEnabled ? { 'audit-logs': auditLogsController } : {}),
+      ...(isAIEnabled ? { ai: aiController } : {}),
+    },
+    routes: {
+      ...eeAdmin.routes,
+      ...(isAuditLogsEnabled ? { 'audit-logs': auditLogsRoutes } : {}),
+      ...(isAIEnabled ? { ai: aiRoutes } : {}),
+    },
+    async register({ strapi }: { strapi: Core.Strapi }) {
+      // Run the the default registration
+      await eeAdmin.register({ strapi });
+      if (isAuditLogsEnabled) {
         // Register an internal audit logs service
         strapi.add('audit-logs', createAuditLogsService(strapi));
         // Register an internal audit logs lifecycle service
@@ -51,15 +57,15 @@ const getAdminEE = () => {
         strapi.add('audit-logs-lifecycle', auditLogsLifecycle);
 
         await auditLogsLifecycle.register();
-      },
-      async destroy({ strapi }: { strapi: Core.Strapi }) {
+      }
+    },
+    async destroy({ strapi }: { strapi: Core.Strapi }) {
+      if (isAuditLogsEnabled) {
         strapi.get('audit-logs-lifecycle').destroy();
-        await eeAdmin.destroy({ strapi });
-      },
-    };
-  }
-
-  return eeAdmin;
+      }
+      await eeAdmin.destroy({ strapi });
+    },
+  };
 };
 
 export default getAdminEE;
