--- packages/core/admin/ee/server/src/routes/ai.ts
@@ -1,14 +1,0 @@
-export default {
-  type: 'admin',
-  routes: [
-    // ai data usage
-    {
-      method: 'GET',
-      path: '/ai-usage',
-      handler: 'admin.getAiUsage',
-      config: {
-        policies: ['admin::isAuthenticatedAdmin'],
-      },
-    },
-  ],
-};
