+++ packages/core/admin/ee/server/src/routes/index.ts
@@ -1,9 +1,7 @@
 import sso from './sso';
 import licenseLimit from './license-limit';
-import ai from './ai';
 
 export default {
   sso,
   'license-limit': licenseLimit,
-  ai,
 };
