+++ packages/core/admin/shared/contracts/users.ts
@@ -45,7 +45,7 @@ export declare namespace UpdateMe {
 }
 
 /**
- * GET /users/me/getAiToken - Get an AI token for the current admin user
+ * GET /getAiToken - Get an AI token for the current admin user
  */
 export declare namespace GetAiToken {
   export interface Request {
@@ -73,24 +73,6 @@ export declare namespace GetOwnPermissions {
 
   export interface Response {
     data: Permission[];
-    error?: errors.ApplicationError;
-  }
-}
-
-/**
- * GET /users/me/ai-token - Get AI token for the current admin user
- */
-export declare namespace GetAiToken {
-  export interface Request {
-    query: {};
-    body: {};
-  }
-
-  export interface Response {
-    data: {
-      token: string;
-      expiresAt?: string;
-    };
     error?: errors.ApplicationError;
   }
 }
